+++ src/app/controllers/curso_especializacao/turmadisciplinaespecializacaoController.js
@@ -8,9 +8,9 @@ import MinistranteMinistraTurmaEspecializacao from '../../models/curso_especiali
 import DocenteMinistrante from '../../models/usuarios/docenteministrante'
 
 // Utils
-import MESSAGES from '../../utils/messages/messages_pt'
-import httpStatus from '../../utils/httpStatus/httpStatus'
-import CustomError from '../../utils/CustomError/CustomError'
+import MESSAGES from '../../utils/response/messages/messages_pt'
+import httpStatus from '../../utils/response/httpStatus/httpStatus'
+import CustomError from '../../utils/response/CustomError/CustomError'
 
 class turmaDisciplinaEspecializacaoController {
