+++ src/app/controllers/instituicao/instituicaoEnsinoController.js
@@ -7,9 +7,9 @@ import InstituicaoEnsinoBrasileira from '../../models/instituicao/instituicaoens
 
 // Utils
 import { Op } from 'sequelize'
-import MESSAGES from '../../utils/messages/messages_pt'
-import CustomError from '../../utils/CustomError/CustomError'
-import httpStatus from '../../utils/httpStatus/httpStatus'
+import MESSAGES from '../../utils/response/messages/messages_pt'
+import CustomError from '../../utils/response/CustomError/CustomError'
+import httpStatus from '../../utils/response/httpStatus/httpStatus'
 
 class instituicaoEnsinoController {
     async post(req, res, brasileira){
