+++ src/app/controllers/ofertacoletiva/turmaOCController.js
@@ -5,9 +5,9 @@ import ProfessorIsF from '../../models/usuarios/professorisf'
 
 // Utils
 import UserTypes from "../../utils/userType/userTypes"
-import httpStatus from "../../utils/httpStatus/httpStatus"
-import MESSAGES from "../../utils/messages/messages_pt"
-import CustomError from "../../utils/CustomError/CustomError"
+import httpStatus from "../../utils/response/httpStatus/httpStatus"
+import MESSAGES from "../../utils/response/messages/messages_pt"
+import CustomError from "../../utils/response/CustomError/CustomError"
 
 class turmaOCController {
     static async verifyExistingClass(name) {
