+++ src/app/models/usuarios/alunoisf.js
@@ -20,7 +20,7 @@ class AlunoIsF extends Model {
             {
                 sequelize,
                 timestamps: false,
-                tableName: 'alunoisf',
+                tableName: 'aluno_isf',
                 indexes: [{
                     name: "primary_key",
                     unique: true,
